pyproject.toml
[tool.poetry]
name = "spare-parts-accounting"
version = "0.1.0"
description = ""
authors = ["roman <imationnn@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
tenacity = "^8.2.3"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"
redis = "^5.0.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "app"]
asyncio_mode = "auto"
